show card data
